backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-cors-headers
python-jose
django-otp
qrcode
Pillow
requests
dj-database-url
